Move native Entity functions out of ScriptEntity into EntityImpl
