+++ include/Engine/Bytecode/ScriptEntity.h
@@ -59,33 +59,6 @@ public:
 	void Remove();
 	void Dispose();
 	bool IsValid();
-	static VMValue VM_SetAnimation(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_ResetAnimation(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_Animate(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_GetIDWithinClass(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_AddToRegistry(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_IsInRegistry(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_RemoveFromRegistry(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_ApplyMotion(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_InView(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_CollidedWithObject(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_GetHitboxFromSprite(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_ReturnHitbox(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_CollideWithObject(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_SolidCollideWithObject(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_TopSolidCollideWithObject(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_ApplyPhysics(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_PropertyExists(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_PropertyGet(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_SetViewVisibility(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_SetViewOverride(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_AddToDrawGroup(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_IsInDrawGroup(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_RemoveFromDrawGroup(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_PlaySound(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_LoopSound(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_StopSound(int argCount, VMValue* args, Uint32 threadID);
-	static VMValue VM_StopAllSounds(int argCount, VMValue* args, Uint32 threadID);
 };
 
 #endif /* ENGINE_BYTECODE_SCRIPTENTITY_H */
